Create stat block abstraction and use
