--- src/app/dashboard/_components/CumulativeScore.tsx
@@ -1,7 +1,0 @@
-export default function CumulativeScore({
-  cumulativeScore,
-}: {
-  cumulativeScore: number;
-}) {
-  return <div>{cumulativeScore}</div>;
-}
