+++ src/app/dashboard/page.tsx
@@ -3,9 +3,8 @@ import {
   getHighestAndLowestScore,
   getCumulativeScore,
 } from "@/server/queries";
-import BattingAverage from "./_components/BattingAverage";
-import HighLowScore from "./_components/HighLowScore";
-import CumulativeScore from "./_components/CumulativeScore";
+
+import BasicStatBlock from "@/components/BasicStatBlock";
 
 export default async function Dashboard() {
   const battingAverage = await getPlayerBattingAverage();
@@ -15,16 +14,38 @@ export default async function Dashboard() {
   return (
     <section className="border-zinc-500 p-5">
       <div className="mb-4">
-        <BattingAverage battingAverage={battingAverage} />
-      </div>
-      <div className="">
-        <HighLowScore
-          highest={highest?.score ?? null}
-          lowest={lowest?.score ?? null}
+        <BasicStatBlock
+          label="Batting Average"
+          value={battingAverage.battingAverage}
+          details={
+            <div>
+              <div className="flex items-center justify-between">
+                <div className="text-base text-gray-400">Rounds Won</div>
+                <div className="text-base font-medium">
+                  {battingAverage.totalHandsWon}
+                </div>
+              </div>
+              <div className="flex items-center justify-between">
+                <div className="text-base text-gray-400">Rounds Played</div>
+                <div className="text-base font-medium">
+                  {battingAverage.totalHandsPlayed}
+                </div>
+              </div>
+            </div>
+          }
         />
       </div>
-      <div className="">
-        <CumulativeScore cumulativeScore={cumulativeScore} />
+      <div className="mb-4">
+        <BasicStatBlock
+          label="High / Low Singe Hand"
+          value={`${highest?.score ?? null} / ${lowest?.score ?? null}`}
+        />
+      </div>
+      <div className="mb-4">
+        <BasicStatBlock
+          label="Total Cumulative Score"
+          value={cumulativeScore.toString()}
+        />
       </div>
     </section>
   );
